refactor: Refactor verify_service
