+++ messenger/controller.py
@@ -1,10 +1,10 @@
 from flask import Blueprint, request
 
-from messenger.service import verify
+from messenger.verify.service import verify_service
 
 messenger_controller = Blueprint('messenger', __name__)
 
 @messenger_controller.get('/')
 def verify_controller():
     query = request.args
-    return verify.verify_service(query)
+    return verify_service(query)
